+++ src/components/layout/content/schulte/SchulteTable.tsx
@@ -1,20 +1,39 @@
-import { useMemo, useState } from 'react'
+import { useState } from 'react'
 import { getCellClasses, shuffle } from '../../../../utils/content.utils'
+import { Button } from '@heroui/react'
+import Confetti from 'react-confetti'
 
 export default function SchulteCounter() {
+	// TODOS: Add final card
+	// TODOS: Add restart button
+	// TODOS: Add click sounds
+
 	const gridSize = 5
 	const limitNumber = 25
-	const cells = useMemo(() => {
+	const [cells, setCells] = useState(() => {
 		const arr = Array.from({ length: gridSize * gridSize }, (_, i) => i + 1)
 		return shuffle(arr)
-	}, [gridSize])
+	})
 
 	const [currentNumber, setCurrentNumber] = useState<number>(1)
+	const [isFinished, setIsFinished] = useState(false)
+
+	const handleRestart = () => {
+		// reset number
+		setCurrentNumber(1)
+		setIsFinished(false)
+
+		// shuffle new grid
+		const arr = Array.from({ length: gridSize * gridSize }, (_, i) => i + 1)
+		setCells(shuffle(arr))
+	}
 
 	return (
 		<div className="flex flex-col justify-start pt-4 items-center w-full h-full p-4 gap-4">
-			<div className="flex flex-col items-center justify-center text-3xl uppercase">
-				<span>Current number:</span> <span className="text-[4rem] font-bold">{currentNumber}</span>{' '}
+			{isFinished && <Confetti width={window.innerWidth} height={window.innerHeight} />}
+			<div className="w-full max-w-[70vmin] flex flex-col items-center justify-center">
+				<span className="text-3xl uppercase">Current number:</span>
+				<span className="text-[4rem] font-bold">{currentNumber}</span>
 			</div>
 			<div className="w-full h-full max-w-[70vmin] max-h-[70vmin] border-4 border-[#fafafa] bg-[#fafafa] rounded-[1.8rem] grid grid-cols-5 grid-rows-5 gap-1 overflow-hidden">
 				{cells.map((num, i) => (
@@ -27,6 +46,7 @@ export default function SchulteCounter() {
 						onClick={() => {
 							if (num === limitNumber && currentNumber === limitNumber) {
 								//TODO: Add custom modal to show results
+								setIsFinished(true)
 								console.log('Table finished')
 							} else if (num === currentNumber) {
 								setCurrentNumber(currentNumber + 1)
@@ -37,6 +57,9 @@ export default function SchulteCounter() {
 					</div>
 				))}
 			</div>
+			<Button variant="light" size="lg" radius="lg" className="text-[#fafafa] font-bold" onPress={handleRestart}>
+				Restart
+			</Button>
 		</div>
 	)
 }